MultiBinder doesn't work well with hierarchical extension points.
This is because it wants to inject Set<X> for every extension point X,
and so it ends up trying to inject Set<X> both in the parent and the
child, which is an error.

So we should just instead concentrate on defining bindings that doesn't
collide, then define a mechanism to externally walk Injector and
list up extension points.

It is a shame that Guice doesn't allow applications and other modules
to do injection-point aware injcetions --- this capability is reserved
for a few built-in bindings, such as Logger.

Originally-Committed-As: ae8c3d7176633d342453b1ba43b83cabeb0d6a58